+++ moncic/ci.py
@@ -1,11 +1,13 @@
 from __future__ import annotations
 import argparse
+from collections import defaultdict
 import contextlib
 import csv
 import logging
 import os
 import shlex
 import shutil
+import stat
 import subprocess
 import sys
 import tempfile
@@ -428,3 +430,47 @@ class Distros(MoncicCommand):
             for info in family.list_distros():
                 output.add_row((info.name, ", ".join(info.shortcuts)))
         output.flush()
+
+
+class Dedup(MoncicCommand):
+    """
+    Deduplicate disk usage in image directories
+    """
+    def run(self):
+        from .btrfs import do_dedupe
+
+        imagedir = self.moncic.config.imagedir
+
+        by_name_size = defaultdict(list)
+        for entry in os.scandir(imagedir):
+            if entry.name.startswith("."):
+                continue
+            if not entry.is_dir():
+                continue
+
+            path = os.path.join(imagedir, entry.name)
+            for (dirpath, dirnames, filenames, dirfd) in os.fwalk(path):
+                relpath = os.path.relpath(dirpath, path)
+                for fn in filenames:
+                    st = os.lstat(fn, dir_fd=dirfd)
+                    if not stat.S_ISREG(st.st_mode):
+                        continue
+                    size = st.st_size
+                    by_name_size[(os.path.join(relpath, fn), size)].append(entry.name)
+
+        with self.moncic.privs.root():
+            total_saved = 0
+            for (name, size), images in by_name_size.items():
+                if len(images) < 2:
+                    continue
+                saved = 0
+                for imgname in images[1:]:
+                    saved += do_dedupe(
+                            os.path.join(imagedir, images[0], name),
+                            os.path.join(imagedir, imgname, name),
+                            size)
+                if saved > 0:
+                    log.info("%s: found in %s, recovered %db", name, ", ".join(images), saved)
+                total_saved += saved
+
+        log.info("Deduplication recovered %db in total", total_saved)