+++ NEWS.md
@@ -1,3 +1,8 @@
+# Version NEXT
+
+* Implemented simple deduplication of files with the same name and size across
+  OS images (#19)
+
 # Version 0.1
 
 * First release